requirements.txt

fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.35
alembic==1.14.0
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.20
email-validator==2.3.0